Move chunked dataset classes into new io.jhdf.dataset.chunked package
