+++ jhdf/src/main/java/io/jhdf/dataset/DatasetLoader.java
@@ -13,8 +13,9 @@ import io.jhdf.HdfFileChannel;
 import io.jhdf.ObjectHeader;
 import io.jhdf.api.Dataset;
 import io.jhdf.api.Group;
+import io.jhdf.dataset.chunked.ChunkedDatasetV3;
+import io.jhdf.dataset.chunked.ChunkedDatasetV4;
 import io.jhdf.exceptions.HdfException;
-import io.jhdf.exceptions.UnsupportedHdfException;
 import io.jhdf.object.message.DataLayoutMessage;
 import io.jhdf.object.message.DataLayoutMessage.ChunkedDataLayoutMessageV3;
 import io.jhdf.object.message.DataLayoutMessage.ChunkedDataLayoutMessageV4;
