+++ jhdf/src/main/java/io/jhdf/dataset/chunked/package-info.java
@@ -1,0 +1,17 @@
+/*******************************************************************************
+ * This file is part of jHDF. A pure Java library for accessing HDF5 files.
+ *
+ * http://jhdf.io
+ *
+ * Copyright 2019 James Mudd
+ *
+ * MIT License see 'LICENSE' file
+ ******************************************************************************/
+/**
+ * Chunked dataset classes.
+ *
+ * These are the classes used when accessing chunked HDF5 datasets.
+ *
+ * @author James Mudd
+ */
+package io.jhdf.dataset.chunked;
